Split adapters into two files; relational and mongo
